rename reporter.http to avoid conflicts w/ lazy loaded workzeug pkgs.
